% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x2y.r
\name{calc_misclass_reduction}
\alias{calc_misclass_reduction}
\title{FUNCTION_TITLE}
\usage{
calc_misclass_reduction(y_hat, y_actual)
}
\arguments{
\item{y_hat}{PARAM_DESCRIPTION}

\item{y_actual}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
