% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x2y.r
\name{dx2y}
\alias{dx2y}
\title{FUNCTION_TITLE}
\usage{
dx2y(d, target = NA, confidence = FALSE)
}
\arguments{
\item{d}{PARAM_DESCRIPTION}

\item{target}{PARAM_DESCRIPTION, Default: NA}

\item{confidence}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
