% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hs3D_knn_functions.r
\name{hs_knn_pred}
\alias{hs_knn_pred}
\title{Make prediction by IDW or prepare neighboring data: observations-distances}
\usage{
hs_knn_pred(
  soil.fun,
  obs.data,
  pred_depth,
  depth_th,
  n,
  p,
  output = list("prediction", "preparation")
)
}
\arguments{
\item{soil.fun}{formula object that defines the relations target/predictors}

\item{obs.data}{training data in the form of tidy soil data}

\item{pred_depth}{prediction depth in cm}

\item{depth_th}{depth treshold}

\item{n}{number of observations used for idw}

\item{p}{power of dist}

\item{output}{Output, Default: list("prediction", "preparation")}
}
\value{
Prediction or data frame
}
\description{
Make prediction by IDW or prepare neighboring data: observations-distances
}
