% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil3D_knn_function.r
\name{ncv.soil3Dknn}
\alias{ncv.soil3Dknn}
\title{ncv.soil3Dknn}
\usage{
ncv.soil3Dknn(soil.fun, obs.data, params, folds, parallel = FALSE, n.cores = 1)
}
\arguments{
\item{soil.fun}{formula object that defines target and predictor variables which will be used in searching for neighboring profiles. (eg. \code{OC~x + y + DEM + TWI + depth}. In this case, \code{x}, \code{y}, \code{DEM} and \code{TWI} are included in the searching algorithm. \code{depth} is mandatory for 3D soil mapping.}

\item{obs.data}{Observation data. Tidy \code{data.frame} with the following columns: \code{ID}, \code{Top}, \code{Bottom}, \verb{Target variable} column, \code{Covariates} columns, \code{x}, \code{y}, \code{depth}.}

\item{params}{\code{data.frame} of possible values of parameters (\code{n.obs}, \code{depth_th}, \code{p}).}

\item{folds}{data.frame with two columns indicating which profiles belong to each fold in outer and inner loop of the nested cross-validation.}

\item{parallel}{logical, if TRUE the computation will perform in parallel, Default: FALSE}

\item{n.cores}{Number of cores for parallel computation, Default: 1}

\item{n.obs}{number of neighboring profiles.Tuning parameter.}
}
\value{
Accuracy measures
}
\description{
performs the nested cross-validation procedure for assessing prediction accuracy.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
