% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil3D_knn_function.r
\name{partitioning}
\alias{partitioning}
\title{FUNCTION_TITLE}
\usage{
partitioning(
  partitioning.data,
  num.folds,
  princ.comp,
  num.means,
  spatial.cluster = TRUE,
  seed = 46,
  cum.prop = 0.9
)
}
\arguments{
\item{partitioning.data}{PARAM_DESCRIPTION}

\item{num.folds}{PARAM_DESCRIPTION}

\item{princ.comp}{PARAM_DESCRIPTION}

\item{num.means}{PARAM_DESCRIPTION}

\item{spatial.cluster}{PARAM_DESCRIPTION, Default: TRUE}

\item{seed}{PARAM_DESCRIPTION, Default: 46}

\item{cum.prop}{PARAM_DESCRIPTION, Default: 0.9}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
