% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil3D_knn_function.r
\name{soil3D_knn}
\alias{soil3D_knn}
\title{soil3D_knn}
\usage{
soil3D_knn(
  soil.fun,
  obs.data,
  pred.data,
  n.obs,
  depth_th,
  p = 1,
  output = "prediction"
)
}
\arguments{
\item{soil.fun}{formula object that defines target and predictor variables which will be used in searching for neighboring profiles. (eg. \code{OC~x + y + DEM + TWI + depth}. In this case, \code{x}, \code{y}, \code{DEM} and \code{TWI} are included in the searching algorithm. \code{depth} is mandatory for 3D soil mapping.}

\item{obs.data}{Observation data. Tidy \code{data.frame} with the following columns: \code{ID}, \code{Top}, \code{Bottom}, \verb{Target variable} column, \code{Covariates} columns, \code{x}, \code{y}, \code{depth}.}

\item{pred.data}{Prediction data. Tidy \code{data.frame} with the following columns: \code{ID}, \code{Covariates} columns, \code{x}, \code{y}, \code{depth}.}

\item{n.obs}{number of neighboring profiles.Tuning parameter.}

\item{depth_th}{depth treshold. Serves to define prediction depth interval: \verb{pdi = pred.data$depth +/- depth_th}, Tuning parameter.}

\item{p}{power of distance. Tuning parameter. Default: 1}

\item{output}{prediction or preparation. If output = "prediction" prediction will be applied, Default: 'prediction'}
}
\value{
prediction based on IDW and gower distance.
}
\description{
Perform interpolation of soil variables in 3D by usin knn approach.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
