% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil3D_knn_function.r
\name{stratpart}
\alias{stratpart}
\title{stratpart}
\usage{
stratpart(
  target.name,
  sp.data,
  num.folds = 5,
  princ.comp = TRUE,
  kmean.vars = NA,
  num.means = 3,
  spatial.cluster = TRUE,
  nested = TRUE,
  seed = 45,
  cum.prop = 0.9
)
}
\arguments{
\item{target.name}{names of target variables}

\item{sp.data}{"SpatialPointsDataFrame" object of data}

\item{num.folds}{Number of folds, Default: 5}

\item{princ.comp}{logical, do the principal compoment shoud be performed, Default: TRUE}

\item{kmean.vars}{Names of variables used for k-means clustering, Default: NA}

\item{num.means}{Number of clusters, Default: 3}

\item{spatial.cluster}{logical, if spatial clustering must be done, Default: TRUE}

\item{nested}{logical, if two step partitioning (outer and inner) must be done, Default: TRUE}

\item{seed}{random number generator seed, Default: 45}

\item{cum.prop}{proportion treshold for principal component analysis, Default: 0.9}
}
\value{
vector of length sp.data, with the numbers indicating which point belongs to each fold. In case of nested = TRUE, the output is data.frame with two vectors, one for outer and one for inner partitioning.
}
\description{
perform n-fold stratified partitioning of data
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
