% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil3D_knn_function.r
\name{tune.soil3Dknn}
\alias{tune.soil3Dknn}
\title{tune.soil3Dknn}
\usage{
tune.soil3Dknn(
  soil.fun,
  obs.data,
  params,
  folds,
  weights = FALSE,
  parallel = FALSE,
  n.cores = 1
)
}
\arguments{
\item{soil.fun}{formula object that defines target and predictor variables which will be used in searching for neighboring profiles. (eg. \code{OC~x + y + DEM + TWI + depth}. In this case, \code{x}, \code{y}, \code{DEM} and \code{TWI} are included in the searching algorithm. \code{depth} is mandatory for 3D soil mapping.}

\item{obs.data}{Observation data. Tidy \code{data.frame} with the following columns: \code{ID}, \code{Top}, \code{Bottom}, \verb{Target variable} column, \code{Covariates} columns, \code{x}, \code{y}, \code{depth}.}

\item{params}{\code{data.frame} of possible values of tuning parameters (\code{depth.th}, \code{n.obs}, \code{p}, \code{d}).}

\item{folds}{vector indicating which profiles belong to each fold.}

\item{weights}{logical, if TRUE weights for gower distance components will be computed}

\item{parallel}{logical, if TRUE the computation will perform in parallel, Default: FALSE}

\item{n.cores}{Number of cores for parallel computation, Default: 1}
}
\value{
\code{params} data.frame with the corresponding accuracy measures.
}
\description{
performs the cross-validation based tuning of meta-parameters.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
