% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil3D_knn_function.r
\name{tune_soil3D_knn}
\alias{tune_soil3D_knn}
\title{tune_soil3D_knn}
\usage{
tune_soil3D_knn(soil.fun, obs.data, params, folds)
}
\arguments{
\item{soil.fun}{formula object that defines target and predictor variables which will be used in searching for neighboring profiles. (eg. \code{OC~x + y + DEM + TWI + depth}. In this case, \code{x}, \code{y}, \code{DEM} and \code{TWI} are included in the searching algorithm. \code{depth} is mandatory for 3D soil mapping.}

\item{obs.data}{Observation data. Tidy \code{data.frame} with the following columns: \code{ID}, \code{Top}, \code{Bottom}, \verb{Target variable} column, \code{Covariates} columns, \code{x}, \code{y}, \code{depth}.}

\item{params}{\code{data.frame} of possible values of parameters (\code{n.obs}, \code{depth_th}, \code{p}).}

\item{folds}{vector indicating which profiles belong to each fold.}

\item{n.obs}{number of neighboring profiles.Tuning parameter.}
}
\value{
params data.frame with the corresponding accuracy measures
}
\description{
performs the cross-validation based tuning of meta-parameters.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
