% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x2y.r
\name{x2y}
\alias{x2y}
\title{FUNCTION_TITLE}
\usage{
x2y(x, y, confidence = FALSE)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{y}{PARAM_DESCRIPTION}

\item{confidence}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
