% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x2y.r
\name{x2y_inner}
\alias{x2y_inner}
\title{FUNCTION_TITLE}
\usage{
x2y_inner(x, y)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{y}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
