% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hs3D_knn_functions.r
\name{soil3Dknn}
\alias{soil3Dknn}
\title{soil 3D knn}
\usage{
soil3Dknn(
  soil.fun,
  pred.data,
  obs.data,
  depth_th,
  n,
  p = 2,
  min_obs = 10,
  radius = NA,
  output = list("prediction", "preparation")
)
}
\arguments{
\item{soil.fun}{formula of soil target/predictors relations}

\item{pred.data}{Prediction data}

\item{obs.data}{Observation data}

\item{depth_th}{depth treshold}

\item{n}{Number of observations}

\item{p}{Power of distance in idw, Default: 2}

\item{min_obs}{Minimum observations at particular prediction depth, Default: 10}

\item{radius}{Searching radius. , Default: NA}

\item{output}{PARAM_DESCRIPTION, Default: list("prediction", "preparation")}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Perform soil 3D knn
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[sf]{st_as_sf}}
\code{\link[purrr]{map}},\code{\link[purrr]{map2}}
\code{\link[nngeo]{st_nn}}
\code{\link[dplyr]{distinct}},\code{\link[dplyr]{filter}},\code{\link[dplyr]{group_split}},\code{\link[dplyr]{mutate}}
\code{\link[tibble]{rownames}}
}
