% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil3D_knn_function.r
\name{soil3Dknn}
\alias{soil3Dknn}
\title{soil3Dknn}
\usage{
soil3Dknn(
  soil.fun,
  obs.data,
  pred.data,
  n.obs,
  depth_th = 10,
  p,
  d,
  output = "prediction",
  parallel = FALSE,
  n.cores = 1
)
}
\arguments{
\item{soil.fun}{soil.fun formula object that defines target and predictor variables which will be used in searching for neighboring profiles. (eg. \code{OC~x + y + DEM + TWI + depth}. In this case, \code{x}, \code{y}, \code{DEM} and \code{TWI} are included in the searching algorithm. \code{depth} is mandatory for 3D soil mapping.}

\item{obs.data}{Observation data. Tidy \code{data.frame} with the following columns: \code{ID}, \code{Top}, \code{Bottom}, \verb{Target variable} column, \code{Covariates} columns, \code{x}, \code{y}, \code{depth}.}

\item{pred.data}{Prediction data. Tidy \code{data.frame} with the following columns: \code{ID}, \code{Covariates} columns, \code{x}, \code{y}, \code{depth}.}

\item{n.obs}{number of neighboring profiles.Tuning parameter.}

\item{depth_th}{depth threshold. Serves to define prediction depth interval: \verb{pdi = pred.data$depth +/- depth_th}, Tuning parameter, Default: 10}

\item{p}{power of horizontal distance. Tuning parameter. Default: 1}

\item{d}{power of vertical distance. Tuning parameter. Default: 1}

\item{output}{prediction or preparation. If output = "prediction" prediction will be applied, Default: 'prediction'}

\item{parallel}{logical, if TRUE the computation will perform in parallel, Default: FALSE}

\item{n.cores}{Number of cores for parallel computation, Default: 1}
}
\value{
prediction based on IDW and gower distance.
}
\description{
soil3Dknn interpolation
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
